Deleted InternalMiiManager
